+++ sentry/src/main/java/io/sentry/Scopes.java
@@ -1245,6 +1245,56 @@ public final class Scopes implements IScopes {
   }
 
   @Override
+  public void setAttribute(final @Nullable String key, final @Nullable Object value) {
+    if (!isEnabled()) {
+      getOptions()
+          .getLogger()
+          .log(
+              SentryLevel.WARNING, "Instance is disabled and this 'setAttribute' call is a no-op.");
+    } else {
+      getCombinedScopeView().setAttribute(key, value);
+    }
+  }
+
+  @Override
+  public void setAttribute(final @Nullable SentryAttribute attribute) {
+    if (!isEnabled()) {
+      getOptions()
+          .getLogger()
+          .log(
+              SentryLevel.WARNING, "Instance is disabled and this 'setAttribute' call is a no-op.");
+    } else {
+      getCombinedScopeView().setAttribute(attribute);
+    }
+  }
+
+  @Override
+  public void setAttributes(final @Nullable SentryAttributes attributes) {
+    if (!isEnabled()) {
+      getOptions()
+          .getLogger()
+          .log(
+              SentryLevel.WARNING,
+              "Instance is disabled and this 'setAttributes' call is a no-op.");
+    } else {
+      getCombinedScopeView().setAttributes(attributes);
+    }
+  }
+
+  @Override
+  public void removeAttribute(final @Nullable String key) {
+    if (!isEnabled()) {
+      getOptions()
+          .getLogger()
+          .log(
+              SentryLevel.WARNING,
+              "Instance is disabled and this 'removeAttribute' call is a no-op.");
+    } else {
+      getCombinedScopeView().removeAttribute(key);
+    }
+  }
+
+  @Override
   public void addFeatureFlag(final @Nullable String flag, final @Nullable Boolean result) {
     combinedScope.addFeatureFlag(flag, result);
   }